Update Edge browser detection
The new user agent added in the test is a Samsung Android device, maybe
running with a Edge webview.

The EdgW/1.0 bit add the end of the UA string confuses the library into
believing that's Edge 1.0.

Whether it's an Edge webview or not, for compatibility reasons it's
better to report the browser as Chrome.